+++ boards/cats_rev1Pro/Core/Inc/FreeRTOSConfig.h
@@ -69,7 +69,7 @@ extern uint32_t SystemCoreClock;
 #define configMINIMAL_STACK_SIZE                ((uint16_t)128)
 #define configTOTAL_HEAP_SIZE                   ((size_t)18 * 1024)
 #define configMAX_TASK_NAME_LEN                 (16)
-#define configUSE_TRACE_FACILITY                1
+#define configUSE_TRACE_FACILITY                0
 #define configUSE_16_BIT_TICKS                  0
 #define configUSE_MUTEXES                       1
 #define configQUEUE_REGISTRY_SIZE               8
